added condition to delete files in Alfresco
git-svn-id: http://svn.bluexml.com/private/S-IDE@11745 c4e9bb49-730a-0410-b52d-bc9a0859fdfe

+++ FrameworksModules/trunk/com.bluexml.side.Framework.alfresco.dataGenerator/src/main/java/com/bluexml/side/Framework/alfresco/dataGenerator/load/ImportACP.java
@@ -75,7 +75,9 @@ public class ImportACP {
 		List<FileInfo> contentFiles = fileFolderService.listFiles(container);
 		if (contentFiles.size() > 0) {
 			for (FileInfo file : contentFiles) {
-				fileFolderService.delete(file.getNodeRef());
+				if (fileFolderService.exists(file.getNodeRef())){
+					fileFolderService.delete(file.getNodeRef());
+				}
 			}
 		}		
 		return container;
